refactor(auth): extract and rename blocking error dialogs
